+++ RELEASE_NOTES.md
@@ -16,6 +16,7 @@ A new header is inserted each time a *tag* is created.
 - gltfio: fix ubershader issues with assignment of dummy textures
 - gltfio: material instances and variants are now accessed via `FilamentInstance` [⚠️ **API Change**]
 - gltfio: the animator can now only be accessed via `FilamentInstance` [⚠️ **API Change**]
+- engine: fix "stable" shadows and make the default cascade splits logarithmic.
 
 ## v1.27.2